navbar styling halfway complete
