+++ src/components/Navbar/Navbar.scss
@@ -1,0 +1,81 @@
+@use "../../partials/typography" as *;
+@use "../../partials/variables" as *;
+
+*,
+*:after,
+*:before {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+section {
+  display: flex;
+  flex-direction: row;
+  border: 1px solid black;
+  padding: 0 2rem 0 2rem;
+}
+.img {
+  height: 11rem;
+  width: 11rem;
+}
+
+.search {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+
+  &__txt {
+    padding-left: 0.25rem;
+  }
+}
+
+.signout {
+  display: flex;
+  align-items: center;
+  border-right: 0.03125rem solid #db0006;
+  padding-right: 0.5rem;
+  padding: 0 1rem 0 1rem;
+  &__txt {
+    padding-left: 0.5rem;
+  }
+}
+
+.main {
+  width: 100%;
+  border: 1px solid blue;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  &__top {
+    width: 100%;
+    height: 5.5rem;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    @include nav;
+    &--left {
+      border: 1px solid red;
+      padding: 1rem;
+      width: 45%;
+      display: flex;
+      flex-direction: row;
+      justify-content: space-between;
+      align-items: center;
+    }
+    &--right {
+      width: 30%;
+      display: flex;
+      flex-direction: row;
+      justify-content: space-between;
+      align-items: center;
+      padding: 0 1rem 0 1rem;
+    }
+  }
+  &__bottom {
+    border: 1px solid magenta;
+    display: flex;
+    flex-direction: row;
+  }
+}
